+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
+## [8.5.0] - 2025-09-01
+### Added
+- Implemented `database` endpoints.
+
 ## [8.4.0] - 2024-10-01
 ### Added
 - Implemented `server/listiso` and `server/mountiso` endpoints.